requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-materializecss-form==1.1.11
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.7
whitenoise==4.1.2
